+++ Sourcery/Models/Enum.swift
@@ -8,17 +8,23 @@ import Foundation
 final class Enum: Type {
     final class Case: NSObject, AutoDiffable {
         final class AssociatedValue: NSObject, AutoDiffable, Typed {
-            let name: String?
+            let localName: String?
+            let externalName: String?
             let typeName: TypeName
 
             /// sourcery: skipEquality
             /// sourcery: skipDescription
             var type: Type?
 
-            init(name: String?, typeName: String, type: Type? = nil) {
-                self.name = name
+            init(localName: String?, externalName: String?, typeName: String, type: Type? = nil) {
+                self.localName = localName
+                self.externalName = externalName
                 self.typeName = TypeName(typeName)
                 self.type = type
+            }
+
+            convenience init(name: String? = nil, typeName: String, type: Type? = nil) {
+                self.init(localName: name, externalName: name, typeName: typeName, type: type)
             }
         }